requirements.txt
numpy
click
python-dateutil
sortedcontainers
pytest